refactor: with updated code
